[#2437] Refactor
