+++ client/src/components/dataset/DatasetTable.jsx
@@ -10,7 +10,7 @@ import DataTableSidebar from './DataTableSidebar';
 import DatasetControls from './DatasetControls';
 import DataTypeContextMenu from './context-menus/DataTypeContextMenu';
 import ColumnContextMenu from './context-menus/ColumnContextMenu';
-import { flowCommonColumnNames, isTransformationColumn, ensurePushIntoArray, datasetHasQuestionGroups } from './../../utilities/utils';
+import { reducerGroup, datasetHasQuestionGroups } from './../../utilities/utils';
 
 require('./DatasetTable.scss');
 
@@ -452,25 +452,9 @@ class DatasetTable extends Component {
       />);
     };
 
-
-    const reducerGroup = (accumulator, c, idx) => {
-      const column = c.set('idx', idx);
-      const groupName = column.get('groupName');
-      const columnName = column.get('columnName');
-      if (groupName === null || groupName === undefined) {
-        if (isTransformationColumn(column)) {
-          return ensurePushIntoArray(accumulator, intTxs, column);
-        } else if (flowCommonColumnNames.has(columnName)) {
-          return ensurePushIntoArray(accumulator, intFormMetadata, column);
-        }
-        return ensurePushIntoArray(accumulator, ' ', column);
-      }
-      return ensurePushIntoArray(accumulator, groupName, column);
-    };
-
-    const groups = columns.reduce(reducerGroup, {});
     let cols;
     if (datasetHasQuestionGroups(columns)) {
+      const groups = columns.reduce(reducerGroup(intFormMetadata, intTxs), {});
       const reducer2 = (accumulator, k, idx) => {
         const columnsGroup = groups[k];
         accumulator.push(
